refactor: modularize community endpoint policy
